// src/app/pages/interfaces/registro.ts
export interface Registro {
  id: number | any;
  tipo: string | any;
  nombre: string | any;
  nombreComercial: string | any;
  dui: string | any;
  nit: string | any;
  nrc: string | any;
  correo: string | any;
  telefono: string | any;
  actividad: string | any;
  pais: string | any;
  departamento: string | any;
  municipio: string | any;
  complemento: string | any;
  campoNuevo: string | any;
}